requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
bert-score==0.3.13
blinker==1.9.0
cachetools==5.5.1
causal-conv1d==1.5.0.post8
certifi==2025.1.31
charset-normalizer==3.4.1
clearml==1.17.2rc0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
crc32c==2.7.1
cycler==0.12.1
databricks-sdk==0.44.0
datasets==3.3.1
deepspeed==0.16.3
Deprecated==1.2.18
dill==0.3.8
docker==7.1.0
docker-pycreds==0.4.0
donfig==0.8.1.post1
einops==0.8.1
evaluate==0.4.3
fastapi==0.115.11
ffmpy==0.5.0
filelock==3.11.0
fire==0.7.0
Flask==3.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
furl==2.1.3
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
gradio==5.20.1
gradio_client==1.7.2
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
imagecodecs==2024.12.30
imageio==2.37.0
importlib_metadata==8.5.0
itk==5.4.2.post1
itk-core==5.4.2.post1
itk-filtering==5.4.2.post1
itk-io==5.4.2.post1
itk-numerics==5.4.2.post1
itk-registration==5.4.2.post1
itk-segmentation==5.4.2.post1
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazy_loader==0.4
lmdb==1.6.2
lpips==0.1.4
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdurl==0.1.2
mlflow==2.20.1
mlflow-skinny==2.20.1
monai==1.4.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nltk==3.9.1
nni==3.0
numcodecs==0.15.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
openslide-python==1.4.1
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.2.1
orderedmultidict==1.0.1
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathlib2==2.3.7.post1
peft==0.14.0
pillow==11.1.0
platformdirs==4.3.6
prettytable==3.14.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyamg==5.2.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydicom==3.0.1
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.9.0
pyngrok==7.2.3
pynrrd==1.1.3
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
PythonWebHDFS==0.2.3
pytorch-ignite==0.4.11
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
responses==0.25.6
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
ruff==0.9.10
safehttpx==0.1.6
safetensors==0.5.2
schema==0.7.7
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.21.0
setproctitle==1.3.4
setuptools==75.8.0
shellingham==1.5.4
SimpleITK==2.4.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.1.10
timm==1.0.14
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.48.3
triton==3.2.0
typeguard==4.1.2
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
Unidecode==1.3.8
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.6
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zarr==3.0.2
zipp==3.21.0
